backend/requirements.txt
asgiref==3.2.3
Django==1.11.28
django-cors-headers==3.0.2
django-filter==1.1.0
djangorestframework==3.9.4
pkg-resources==0.0.0
pytz==2019.3
sqlparse==0.3.0
